[fli-iam#1898] refactor processing/execution service
