+++ shanoir-ng-front/src/app/vip/execution.data-service.ts
@@ -1,0 +1,24 @@
+import { Injectable } from '@angular/core';
+import { BehaviorSubject, Observable } from 'rxjs';
+import { Pipeline } from './models/pipeline';
+
+@Injectable({
+    providedIn: 'root'
+})
+export class ExecutionDataService {
+
+    //observables
+    public selectedDatasets: Set<number>;
+    public selectedPipeline: Pipeline;
+
+    constructor() {
+    }
+
+    public setDatasets(datasetsIds: Set<number>) {
+        this.selectedDatasets = datasetsIds;
+    }
+
+    public setPipeline(pipeline: Pipeline) {
+        this.selectedPipeline = pipeline;
+    }
+}
