+++ shanoir-ng-front/src/app/vip/execution/execution.component.css
