+++ shanoir-ng-front/src/app/vip/pipelines/pipeline/pipeline.component.css
