+++ shanoir-ng-front/src/app/vip/pipelines/pipelines.component.spec.ts
