+++ shanoir-ng-front/src/app/vip/pipelines/pipelines.component.ts
@@ -3,7 +3,7 @@ import { Router } from '@angular/router';
 import { BreadcrumbsService } from 'src/app/breadcrumbs/breadcrumbs.service';
 import { Pipeline } from 'src/app/vip/models/pipeline';
 import { VipClientService } from 'src/app/vip/shared/vip-client.service';
-import { ProcessingService } from '../processing.service';
+import { ExecutionDataService } from '../execution.data-service';
 import {Mode} from "../../shared/components/entity/entity.component.abstract";
 
 @Component({
@@ -17,7 +17,7 @@ export class PipelinesComponent implements OnInit {
   selectedPipeline:Pipeline;
   descriptionLoading:boolean;
 
-  constructor(private breadcrumbsService: BreadcrumbsService, private vipClientService: VipClientService, private router: Router, private processingService:ProcessingService) {
+  constructor(private breadcrumbsService: BreadcrumbsService, private vipClientService: VipClientService, private router: Router, private processingService:ExecutionDataService) {
     this.pipelines = [];
     this.descriptionLoading = false;
 
@@ -46,16 +46,23 @@ export class PipelinesComponent implements OnInit {
     )
   }
 
+  isSelectedDatasets(): boolean {
+      return this.processingService.selectedDatasets && this.processingService.selectedDatasets.size > 0;
+  }
+
   choosePipeLine(){
     this.processingService.setPipeline(this.selectedPipeline);
     let filesParam = 0;
-    let datsetNumbers = this.processingService.selectedDatasetsValue?.size;
-    // Here we are going to calculate the number of possible executions in parralel
+    // Here we are going to calculate the number of possible executions in parallel
     this.selectedPipeline.parameters?.forEach(parameter => {
         if (parameter.type == 'File') {
             filesParam += 1;
         }
     })
-    this.router.navigate(['processing/execution']);
+    this.router.navigate(['execution']);
   }
+
+  navigateToSolr(): void{
+        this.router.navigate(['/solr-search']);
+    }
 }
